requirements.txt
django
djangorestframework
djangorestframework-simplejwt
django-cors-headers
pillow
django-filter
stripe
razorpay
python-decouple
psycopg2-binary
drf-spectacular
celery
django_redis
qrcode
langgraph
langchain
dotenv
fuzzywuzzy
langchain-core
langchain-perplexity
python-Levenshtein
django-storages
boto3
olefile